[main] deprecated old main page
